moved some code in an extrafile to keep the main_app file not to large.
